Move to a client/server folder structure.
