+++ client/components/entries/index.js
