+++ client/components/icon/index.js
