+++ client/images/icon-512x512.png
