+++ client/js/copy-text/index.js
