+++ client/js/services/entry-service.js
