+++ client/js/services/user-service.js
