+++ client/js/swipe/index.js
