+++ server/middleware/strict-transport-security.js
