+++ server/models/entry-model.js
