+++ server/services/entry-service.js
